requirements.txt
glfw 
pyopengl
pyrr
Pillow
PyOpenGL_accelerate